// crm-risk-assessment-view/src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Board = {
  __typename?: 'Board';
  /** The board log events. */
  activity_logs: Maybe<Array<Maybe<ActivityLogType>>>;
  /** The board's folder unique identifier. */
  board_folder_id: Maybe<Scalars['ID']['output']>;
  /** The board's kind (public / private / share). */
  board_kind: BoardKind;
  /** The board's collaborators */
  collaborators: Array<Maybe<User>>;
  /** The board's visible columns. */
  columns: Maybe<Array<Maybe<Column>>>;
  /** Get the board communication value - typically meeting ID */
  communication: Maybe<Scalars['JSON']['output']>;
  /** The creator of the board. */
  creator: User;
  /** The board's description. */
  description: Maybe<Scalars['String']['output']>;
  /** The board's visible groups. */
  groups: Maybe<Array<Maybe<Group>>>;
  /** The unique identifier of the board. */
  id: Scalars['ID']['output'];
  /** The Board's item nickname, one of a predefined set of values, or a custom user value */
  item_terminology: Maybe<Scalars['String']['output']>;
  /** The number of items on the board */
  items_count: Maybe<Scalars['Int']['output']>;
  /** The board's items (rows). */
  items_page: ItemsResponse;
  /** The board's name. */
  name: Scalars['String']['output'];
  /**
   * The owner of the board.
   * @deprecated This field returned creator of the board. Please use 'creator' or 'owners' fields instead.
   */
  owner: User;
  /** List of user board owners */
  owners: Array<Maybe<User>>;
  /** The board's permissions. */
  permissions: Scalars['String']['output'];
  /** The board's state (all / active / archived / deleted). */
  state: State;
  /** The board's subscribers. */
  subscribers: Array<Maybe<User>>;
  /** The board's specific tags. */
  tags: Maybe<Array<Maybe<Tag>>>;
  /** List of team board owners */
  team_owners: Maybe<Array<Team>>;
  /** The board's team subscribers. */
  team_subscribers: Maybe<Array<Team>>;
  /** The top group at this board. */
  top_group: Group;
  /** The board object type. */
  type: Maybe<BoardObjectType>;
  /** The last time the board was updated at. */
  updated_at: Maybe<Scalars['ISO8601DateTime']['output']>;
  /** The board's updates. */
  updates: Maybe<Array<Update>>;
  /** The Board's url */
  url: Scalars['String']['output'];
  /** The board's views. */
  views: Maybe<Array<Maybe<BoardView>>>;
  /** The workspace that contains this board (null for main workspace). */
  workspace: Maybe<Workspace>;
  /** The board's workspace unique identifier (null for main workspace). */
  workspace_id: Maybe<Scalars['ID']['output']>;
};
// ...
export type ButtonValue = ColumnValue & {
  __typename?: 'ButtonValue';
  /** The button's color in hex value. */
  color: Maybe<Scalars['String']['output']>;
  /** The column that this value belongs to. */
  column: Column;
  /** The column's unique identifier. */
  id: Scalars['ID']['output'];
  /** The button's label. */
  label: Maybe<Scalars['String']['output']>;
  text: Maybe<Scalars['String']['output']>;
  /** The column's type. */
  type: ColumnType;
  /** The column's raw value in JSON format. */
  value: Maybe<Scalars['JSON']['output']>;
};
// ...
export type Column = {
  __typename?: 'Column';
  /** Is the column archived or not. */
  archived: Scalars['Boolean']['output'];
  /** The column's description. */
  description: Maybe<Scalars['String']['output']>;
  /** The column's unique identifier. */
  id: Scalars['ID']['output'];
  /** The column's settings in a string form. */
  settings_str: Scalars['String']['output'];
  /** The column's title. */
  title: Scalars['String']['output'];
  /** The column's type. */
  type: ColumnType;
  /** The column's width. */
  width: Maybe<Scalars['Int']['output']>;
};
// ...
export type Complexity = {
  __typename?: 'Complexity';
  /** The remainder of complexity after the query's execution. */
  after: Scalars['Int']['output'];
  /** The remainder of complexity before the query's execution. */
  before: Scalars['Int']['output'];
  /** The specific query's complexity. */
  query: Scalars['Int']['output'];
  /** How long in seconds before the complexity budget is reset */
  reset_in_x_seconds: Scalars['Int']['output'];
};
// ...
export type ItemsResponse = {
  __typename?: 'ItemsResponse';
  /**
   * An opaque cursor that represents the position in the list after the last
   * returned item. Use this cursor for pagination to fetch the next set of items.
   * If the cursor is null, there are no more items to fetch.
   */
  cursor: Maybe<Scalars['String']['output']>;
  /** The items associated with the cursor. */
  items: Array<Item>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Get the connected account's information. */
  account: Maybe<Account>;
  /** Get a collection of installs of an app. */
  app_installs: Maybe<Array<Maybe<AppInstall>>>;
  /** Get the current app subscription. Note: This query does not work in the playground */
  app_subscription: Maybe<Array<Maybe<AppSubscription>>>;
  /** Get operations counter current value */
  app_subscription_operations: Maybe<AppSubscriptionOperationsCounter>;
  app_subscriptions: AppSubscriptions;
  /** Get apps monetization information for an account */
  apps_monetization_info: Maybe<AppsMonetizationInfo>;
  /** Get apps monetization status for an account */
  apps_monetization_status: Maybe<AppMonetizationStatus>;
  /** Get a collection of assets by ids. */
  assets: Maybe<Array<Maybe<Asset>>>;
  /** Get a collection of boards. */
  boards: Maybe<Array<Maybe<Board>>>;
  /** Get the complexity data of your queries. */
  complexity: Maybe<Complexity>;
  custom_activity: Maybe<Array<CustomActivity>>;
  /** Get a collection of docs. */
  docs: Maybe<Array<Maybe<Document>>>;
  /** Get a collection of folders. Note: This query won't return folders from closed workspaces to which you are not subscribed */
  folders: Maybe<Array<Maybe<Folder>>>;
  /** Get a collection of items. */
  items: Maybe<Array<Maybe<Item>>>;
  /** Search items by multiple columns and values. */
  items_page_by_column_values: ItemsResponse;
  marketplace_app_discounts: Array<MarketplaceAppDiscount>;
  /** Get the connected user's information. */
  me: Maybe<User>;
  /** Get next pages of board's items (rows) by cursor. */
  next_items_page: ItemsResponse;
  /** Get a collection of tags. */
  tags: Maybe<Array<Maybe<Tag>>>;
  /** Get a collection of teams. */
  teams: Maybe<Array<Maybe<Team>>>;
  /** Fetches timeline items for a given item */
  timeline: Maybe<TimelineResponse>;
  timeline_item: Maybe<TimelineItem>;
  /** Get a collection of updates. */
  updates: Maybe<Array<Update>>;
  /** Get a collection of users. */
  users: Maybe<Array<Maybe<User>>>;
  /** Get the API version in use */
  version: Version;
  /** Get a list containing the versions of the API */
  versions: Array<Version>;
  /** Get a collection of webhooks for the board */
  webhooks: Maybe<Array<Maybe<Webhook>>>;
  /** Get a collection of workspaces. */
  workspaces: Maybe<Array<Maybe<Workspace>>>;
};
// ...
export type GetBoardItemsQuery = { __typename?: 'Query', complexity: { __typename?: 'Complexity', before: number, after: number, query: number, reset_in_x_seconds: number } | null, boards: Array<{ __typename?: 'Board', id: string, columns: Array<{ __typename?: 'Column', id: string, title: string, type: ColumnType } | null> | null, items_page: { __typename?: 'ItemsResponse', cursor: string | null, items: Array<{ __typename?: 'Item', id: string, name: string, column_values: Array<{ __typename?: 'BoardRelationValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'ButtonValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'CheckboxValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'ColorPickerValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'CountryValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'CreationLogValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'DateValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'DependencyValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'DocValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'DropdownValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'EmailValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'FileValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'FormulaValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'GroupValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'HourValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'IntegrationValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'ItemIdValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'LastUpdatedValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'LinkValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'LocationValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'LongTextValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'MirrorValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'NumbersValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'PeopleValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'PersonValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'PhoneValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'ProgressValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'RatingValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'StatusValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'SubtasksValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'TagsValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'TeamValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'TextValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'TimeTrackingValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'TimelineValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'UnsupportedValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'VoteValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'WeekValue', id: string, text: string | null, value: any | null, type: ColumnType } | { __typename?: 'WorldClockValue', id: string, text: string | null, value: any | null, type: ColumnType }>, subitems: Array<{ __typename?: 'Item', id: string } | null> | null }> } } | null> | null };

export type GetItemTimelineQueryVariables = Exact<{
  itemId: Scalars['ID']['input'];
}>;


export type GetItemTimelineQuery = { __typename?: 'Query', timeline: { __typename?: 'TimelineResponse', timeline_items_page: { __typename?: 'TimelineItemsPage', cursor: string | null, timeline_items: Array<{ __typename?: 'TimelineItem', id: string | null, custom_activity_id: string | null, content: string | null }> } } | null };